Introduce a new fashion to deal with file attributes. Not completly finished now, but upload iw working.

+++ melody.common/src/main/java/com/wat/melody/common/files/LocalFileSystem.java
@@ -15,7 +15,8 @@ import java.nio.file.Paths;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.nio.file.attribute.FileAttribute;
 
-import org.apache.commons.io.FileUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import com.wat.melody.common.files.exception.WrapperAccessDeniedException;
 import com.wat.melody.common.files.exception.WrapperDirectoryNotEmptyException;
@@ -23,6 +24,7 @@ import com.wat.melody.common.files.exception.WrapperFileAlreadyExistsException;
 import com.wat.melody.common.files.exception.WrapperNoSuchFileException;
 import com.wat.melody.common.files.exception.WrapperNotDirectoryException;
 import com.wat.melody.common.files.exception.WrapperNotLinkException;
+import com.wat.melody.common.messages.Msg;
 
 /**
  * 
@@ -30,6 +32,8 @@ import com.wat.melody.common.files.exception.WrapperNotLinkException;
  * 
  */
 public class LocalFileSystem implements FileSystem {
+
+	private static Logger log = LoggerFactory.getLogger(LocalFileSystem.class);
 
 	public LocalFileSystem() {
 	}
@@ -40,21 +44,37 @@ public class LocalFileSystem implements FileSystem {
 
 	@Override
 	public boolean exists(Path path, LinkOption... options) {
+		if (path == null || path.toString().length() == 0) {
+			throw new IllegalArgumentException(path + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
 		return Files.exists(path, options);
 	}
 
 	@Override
 	public boolean isDirectory(Path path, LinkOption... options) {
+		if (path == null || path.toString().length() == 0) {
+			throw new IllegalArgumentException(path + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
 		return Files.isDirectory(path, options);
 	}
 
 	@Override
 	public boolean isRegularFile(Path path, LinkOption... options) {
+		if (path == null || path.toString().length() == 0) {
+			throw new IllegalArgumentException(path + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
 		return Files.isRegularFile(path, options);
 	}
 
 	@Override
 	public boolean isSymbolicLink(Path path) {
+		if (path == null || path.toString().length() == 0) {
+			throw new IllegalArgumentException(path + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
 		return Files.isSymbolicLink(path);
 	}
 
@@ -62,8 +82,12 @@ public class LocalFileSystem implements FileSystem {
 	public void createDirectory(Path dir, FileAttribute<?>... attrs)
 			throws IOException, NoSuchFileException,
 			FileAlreadyExistsException, AccessDeniedException {
+		if (dir == null || dir.toString().length() == 0) {
+			throw new IllegalArgumentException(dir + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
 		try {
-			Files.createDirectory(dir, attrs);
+			Files.createDirectory(dir);
 		} catch (NoSuchFileException Ex) {
 			throw new WrapperNoSuchFileException(Ex.getFile());
 		} catch (FileAlreadyExistsException Ex) {
@@ -73,18 +97,54 @@ public class LocalFileSystem implements FileSystem {
 		} catch (AccessDeniedException Ex) {
 			throw new WrapperAccessDeniedException(Ex.getFile());
 		}
+		setAttributes(dir, attrs);
 	}
 
 	@Override
 	public void createDirectories(Path dir, FileAttribute<?>... attrs)
 			throws IOException, FileAlreadyExistsException,
 			AccessDeniedException {
+		if (dir == null || dir.toString().length() == 0
+				|| dir.getNameCount() < 1) {
+			return;
+		}
 		try {
-			Files.createDirectories(dir, attrs);
+			createDirectory(dir, attrs);
+			return;
+		} catch (NoSuchFileException Ex) {
+			// if the top first dir cannot be created => raise an error
+			if (dir.getNameCount() <= 1) {
+				throw Ex;
+			}
 		} catch (FileAlreadyExistsException Ex) {
-			throw new WrapperFileAlreadyExistsException(Ex.getFile());
-		} catch (AccessDeniedException Ex) {
-			throw new WrapperAccessDeniedException(Ex.getFile());
+			// if the file is a link on a dir or a dir => no error
+			try {
+				if (readAttributes0(dir).isDirectory()) {
+					return;
+				}
+			} catch (NoSuchFileException Exx) {
+				// concurrency pb : should recreate ?
+			}
+			throw Ex;
+		} catch (IOException Ex) {
+			throw Ex;
+		}
+		// if dir cannot be created => create its parent
+		Path parent = null;
+		parent = dir.resolve("..").normalize();
+		createDirectories(parent, attrs);
+		try {
+			createDirectory(dir, attrs);
+		} catch (FileAlreadyExistsException Ex) {
+			// if the file is a link on a dir or a dir => no error
+			try {
+				if (readAttributes0(dir).isDirectory()) {
+					return;
+				}
+			} catch (NoSuchFileException Exx) {
+				// concurrency pb : should recreate ?
+			}
+			throw Ex;
 		}
 	}
 
@@ -92,8 +152,16 @@ public class LocalFileSystem implements FileSystem {
 	public void createSymbolicLink(Path link, Path target,
 			FileAttribute<?>... attrs) throws IOException, NoSuchFileException,
 			FileAlreadyExistsException, AccessDeniedException {
+		if (link == null || link.toString().length() == 0) {
+			throw new IllegalArgumentException(link + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
+		if (target == null || target.toString().length() == 0) {
+			throw new IllegalArgumentException(target + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
 		try {
-			Files.createSymbolicLink(link, target, attrs);
+			Files.createSymbolicLink(link, target);
 		} catch (NoSuchFileException Ex) {
 			throw new WrapperNoSuchFileException(Ex.getFile());
 		} catch (FileAlreadyExistsException Ex) {
@@ -101,11 +169,16 @@ public class LocalFileSystem implements FileSystem {
 		} catch (AccessDeniedException Ex) {
 			throw new WrapperAccessDeniedException(Ex.getFile());
 		}
+		setAttributes(link, attrs);
 	}
 
 	@Override
 	public Path readSymbolicLink(Path link) throws IOException,
 			NoSuchFileException, NotLinkException, AccessDeniedException {
+		if (link == null || link.toString().length() == 0) {
+			throw new IllegalArgumentException(link + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
 		try {
 			// /!\ This will remove the trailing '/'
 			return Paths.get(Files.readSymbolicLink(link).toString());
@@ -121,6 +194,10 @@ public class LocalFileSystem implements FileSystem {
 	@Override
 	public void delete(Path path) throws IOException, NoSuchFileException,
 			DirectoryNotEmptyException, AccessDeniedException {
+		if (path == null || path.toString().length() == 0) {
+			throw new IllegalArgumentException(path + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
 		try {
 			Files.delete(path);
 		} catch (NoSuchFileException Ex) {
@@ -135,6 +212,10 @@ public class LocalFileSystem implements FileSystem {
 	@Override
 	public boolean deleteIfExists(Path path) throws IOException,
 			DirectoryNotEmptyException, AccessDeniedException {
+		if (path == null || path.toString().length() == 0) {
+			throw new IllegalArgumentException(path + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
 		try {
 			return Files.deleteIfExists(path);
 		} catch (DirectoryNotEmptyException Ex) {
@@ -145,11 +226,37 @@ public class LocalFileSystem implements FileSystem {
 	}
 
 	@Override
-	public void deleteDirectory(Path dir) throws IOException {
-		/*
-		 * TODO : exception thrown by this method doesn't respect the contract
-		 */
-		FileUtils.deleteDirectory(dir.toFile());
+	public void deleteDirectory(Path dir) throws IOException,
+			NotDirectoryException, AccessDeniedException {
+		try {
+			for (Path entry : newDirectoryStream(dir)) {
+				if (isDirectory(entry, LinkOption.NOFOLLOW_LINKS)) {
+					deleteDirectory(entry);
+				} else {
+					deleteIfExists(entry);
+				}
+			}
+		} catch (NoSuchFileException ignored) {
+		}
+		deleteIfExists(dir);
+	}
+
+	@Override
+	public DirectoryStream<Path> newDirectoryStream(Path path)
+			throws IOException, NotDirectoryException, NoSuchFileException {
+		if (path == null || path.toString().length() == 0) {
+			throw new IllegalArgumentException(path + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
+		try {
+			return Files.newDirectoryStream(path);
+		} catch (NoSuchFileException Ex) {
+			throw new WrapperNoSuchFileException(Ex.getFile());
+		} catch (NotDirectoryException Ex) {
+			throw new WrapperNotDirectoryException(Ex.getFile());
+		} catch (AccessDeniedException Ex) {
+			throw new WrapperAccessDeniedException(Ex.getFile());
+		}
 	}
 
 	@Override
@@ -166,6 +273,10 @@ public class LocalFileSystem implements FileSystem {
 
 	private EnhancedFileAttributes _readAttributes(Path path)
 			throws IOException, NoSuchFileException, AccessDeniedException {
+		if (path == null || path.toString().length() == 0) {
+			throw new IllegalArgumentException(path + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
 		BasicFileAttributes pathAttrs = readAttributes0(path,
 				LinkOption.NOFOLLOW_LINKS);
 		Path target = null;
@@ -187,16 +298,31 @@ public class LocalFileSystem implements FileSystem {
 	}
 
 	@Override
-	public DirectoryStream<Path> newDirectoryStream(Path path)
-			throws IOException, NotDirectoryException, NoSuchFileException {
-		try {
-			return Files.newDirectoryStream(path);
-		} catch (NoSuchFileException Ex) {
-			throw new WrapperNoSuchFileException(Ex.getFile());
-		} catch (NotDirectoryException Ex) {
-			throw new WrapperNotDirectoryException(Ex.getFile());
-		} catch (AccessDeniedException Ex) {
-			throw new WrapperAccessDeniedException(Ex.getFile());
+	public void setAttributes(Path path, FileAttribute<?>... attributes)
+			throws IOException, NoSuchFileException, AccessDeniedException {
+		if (path == null || path.toString().length() == 0) {
+			throw new IllegalArgumentException(path + ": Not accepted. "
+					+ "Must be a valid " + Path.class.getCanonicalName() + ".");
+		}
+		if (attributes == null) {
+			return;
+		}
+		/*
+		 * TODO: should be wrapped in an IllegalFileAttributeException, with
+		 * responsibilities to the caller to deal with.
+		 */
+		for (FileAttribute<?> attr : attributes) {
+			if (attr == null) {
+				continue;
+			}
+			try {
+				Files.setAttribute(path, attr.name(), attr.value(),
+						LinkOption.NOFOLLOW_LINKS);
+			} catch (UnsupportedOperationException | IllegalArgumentException
+					| ClassCastException Ex) {
+				log.warn(Msg.bind(Messages.LocalFSMsg_SKIP_ATTR, attr, path));
+			}
+
 		}
 	}
